Add '-g' opt for graph output of modified targets
A graph visualisation can be helpful in order to get a better
understanding of the workings of a GNU Make based project.
Particularly if you are debugging why something keeps rebuilding each
pass or if you are very new to a projekt.

All the information that can be gleaned from this graph output is
already present in the output of --print-database. However we strive to
present it more succinct. Only the targets that were modified in the
goal is present and the relationship between targets and prerequisite
is hopefully more clear.

The output from the goal grahps is stored in files called [goal].dot
in the starting directory.

The hope is also that this can be paired with already existing viewers
and explorers, since the dot language is well known (29 years!).

An example (rebuilding GNU Make):
  $ touch src/file.c
  $ AM_MAKEFLAGS="-g" make -g
  Making all in lib
  make[1]: Entering directory '/home/jonas/sandbox/mymake/lib'
  /home/jonas/sandbox/mymake/./make -g all-recursive
  make[2]: Entering directory '/home/jonas/sandbox/mymake/lib'
  make[3]: Entering directory '/home/jonas/sandbox/mymake/lib'
  make[3]: Nothing to be done for 'all-am'.
  make[3]: Leaving directory '/home/jonas/sandbox/mymake/lib'
  make[2]: Writing dependency graph to '/home/jonas/sandbox/mymake/lib/all-recursive.dot'
  make[2]: Leaving directory '/home/jonas/sandbox/mymake/lib'
  make[1]: Writing dependency graph to '/home/jonas/sandbox/mymake/lib/all.dot'
  make[1]: Leaving directory '/home/jonas/sandbox/mymake/lib'
  Making all in po
  make[1]: Entering directory '/home/jonas/sandbox/mymake/po'
  make[1]: Nothing to be done for 'all'.
  make[1]: Leaving directory '/home/jonas/sandbox/mymake/po'
  Making all in doc
  make[1]: Entering directory '/home/jonas/sandbox/mymake/doc'
  make[1]: Nothing to be done for 'all'.
  make[1]: Leaving directory '/home/jonas/sandbox/mymake/doc'
  make[1]: Entering directory '/home/jonas/sandbox/mymake'

  [...]
  [gcc building]
  [...]

  make[1]: Writing dependency graph to '/home/jonas/sandbox/mymake/all-am.dot'
  make[1]: Leaving directory '/home/jonas/sandbox/mymake'
  make: Writing dependency graph to '/home/jonas/sandbox/mymake/all.dot'

  $ cat all-am.dot
  digraph all-am {
  "make" -> "src/file.o"
  "all-am" -> "make"
  "all-am"
  }

  $ cat all.dot
  digraph all {
  "all" -> "all-recursive" [label="forced: PHONY prerequisite"]
  "all"
  }